core/requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
amqp==5.2.0
anyio==4.3.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
billiard==4.2.0
celery==5.3.6
certifi==2024.2.2
cffi==1.16.0
channels==4.0.0
channels-redis==4.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
cron-descriptor==1.4.3
cryptography==42.0.5
daphne==4.1.0
decouple==0.0.7
defusedxml==0.7.1
Django==5.0.3
django-allauth==0.61.1
django-celery-beat==2.6.0
django-environ==0.11.2
django-nextjs==2.4.0
django-timezone-field==6.1.0
djangorestframework==3.15.0
djangorestframework-simplejwt==5.3.1
dj-rest-auth==5.0.2
exceptiongroup==1.2.0
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
iniconfig==2.0.0
kombu==5.3.6
msgpack==1.0.8
multidict==6.0.5
mysqlclient==2.2.4
oauthlib==3.2.2
outcome==1.3.0.post0
packaging==24.0
pillow==10.2.0
pluggy==1.4.0
prompt-toolkit==3.0.43
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pyOpenSSL==24.1.0
PySocks==1.7.1
pytest==8.1.1
pytest-asyncio==0.23.6
pytest-django==4.8.0
python-crontab==3.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==21.1.1
python3-openid==3.2.0
PyYAML==6.0.1
redis==5.0.3
requests==2.31.0
requests-oauthlib==1.4.0
service-identity==24.1.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.4.4
telegram==0.0.1
tomli==2.0.1
trio==0.25.0
trio-websocket==0.11.1
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.28.1
uvloop==0.19.0
vine==5.1.0
watchfiles==0.21.0
wcwidth==0.2.13
websockets==12.0
whitenoise==6.6.0
wsproto==1.2.0
yarl==1.9.4
zope.interface==6.2
drf-yasg
django-cors-headers